move yi-selector to search module
