+++ src/app/elements/card/yi-card.jade
@@ -1,10 +1,10 @@
 template
-  header(class="${typeClass}", click.delegate="router.navigate(report.slug)")
+  header(class="${report.typeText}", click.delegate="router.navigate(report.slug)")
     figure.title
       img(src='http://lorempixel.com/200/90/city')
       figcaption
-        h1 ${report.name}
-        h2 ${category} - ${subCategory}
+        h1 ${report.title}
+        h2 ${report.category.parent.name} - ${report.category.name}
     figure.category
       svg(width='18', height='57')
         path.yi-card-icon(d='M8.6,0.5c4.5,0,8.1,3.6,8.1,8.1v21.5c0,4.5-3.6,8.1-8.1,8.1c-4.5,0-8.1-3.6-8.1-8.1V8.6 C0.5,4.1,4.1,0.5,8.6,0.5z M6.9,56.8V38.2 M10.3,56.8V38.2 M12.4,8.6c0-2.1-1.7-3.8-3.8-3.8c-2.1,0-3.8,1.7-3.8,3.8 c0,2.1,1.7,3.8,3.8,3.8C10.7,12.3,12.4,10.6,12.4,8.6z M4.9,19.4c0,2.1,1.7,3.8,3.8,3.8c2.1,0,3.8-1.7,3.8-3.8 c0-2.1-1.7-3.8-3.8-3.8C6.5,15.6,4.9,17.3,4.9,19.4z M4.9,30.2c0,2.1,1.7,3.8,3.8,3.8s3.8-1.7,3.8-3.8c0-2.1-1.7-3.8-3.8-3.8 S4.9,28.1,4.9,30.2z')
