+++ src/app/reports/reports.jade
@@ -1,8 +1,6 @@
 template
   import(from="app/elements/marker/yi-marker")
   import(from="app/elements/card/yi-card")
-  import(from="app/elements/selector/yi-selector")
-  import(from="app/elements/selector/yi-button")
 
   section#map-area
     core-toolbar
@@ -23,18 +21,6 @@ template
       click.trigger="mrk.toggleType()",
       ref.view-model="mrk"
       )
-
-    core-toolbar.bottom
-      yi-icon-button(icon="map-center")
-      //- create report
-      //- TODO: class interpolation aurelia/binding#33
-      yi-selector(icon="add", select.call="newReport()")
-        // request or complain
-        yi-button(title="${type.name}", value="${type.slug}", class="${type.slug}", repeat.for="type of cats", click.delegate="$parent.selectorType=type.slug")
-          // main categories
-          yi-button(title="${cat.name}", value="${cat.slug}", repeat.for="cat of type.categories")
-            // sub-categories
-            yi-button(title="${subcat.name}", value="${subcat.slug}", repeat.for="subcat of cat.categories")
 
   //- sub routes(search, detail, new)
   router-view(id="${activeSubSection}")
