+++ src/styles/reports.styl
@@ -35,26 +35,51 @@
       position relative
       z-index 1
 
-yi-selector
-  absolute bottom 0 right rightPanelWidth
-  z-index 1
-  margin 0
+  yi-card-list
+    background fade-out(colorBlack, 20%)
+    display flex
+    flex 1
+    flex-direction column
+    flex-shrink 0
+    justify-content space-around
+    height 100%
+    padding 2rem 1rem
+    max-width rightPanelWidth
+    min-width rightPanelWidth
+    overflow-x hidden
+    overflow-y auto
+    z-index 2
 
-  &.opened
-    left 0
-    height 100vh
+    &::-webkit-scrollbar
+      width 4px
+      height 4px
+    &::-webkit-scrollbar-thumb
+      background fade-out(#fff, 50%)
+      border-radius 2px
 
-  @media queryMedium
-    right 0
+    yi-card
+      margin 1rem auto
 
-  &.request
-    &::shadow li yi-button::shadow, yi-button::shadow
-      button, .btn
-        background-color colorBlue
-  &.complain
-    &::shadow li yi-button::shadow, yi-button::shadow
-      button, .btn
-        background-color colorOrange
+  yi-selector
+    absolute bottom 0 right rightPanelWidth
+    z-index 1
+    margin 0
+
+    &.opened
+      left 0
+      height 100vh
+
+    @media queryMedium
+      right 0
+
+    &.request
+      &::shadow li yi-button::shadow, yi-button::shadow
+        button, .btn
+          background-color colorBlue
+    &.complain
+      &::shadow li yi-button::shadow, yi-button::shadow
+        button, .btn
+          background-color colorOrange
 
 yi-button
   &.request::shadow, &.complain::shadow
@@ -63,31 +88,6 @@ yi-button
   &.complain::shadow, &.complain::shadow
     button, .btn
       background-color colorOrange !important
-
-yi-card-list
-  background fade-out(colorBlack, 20%)
-  display flex
-  flex 1
-  flex-direction column
-  flex-shrink 0
-  justify-content space-around
-  height 100%
-  padding 2rem 1rem
-  max-width rightPanelWidth
-  min-width rightPanelWidth
-  overflow-x hidden
-  overflow-y auto
-  z-index 2
-
-  &::-webkit-scrollbar
-    width 4px
-    height 4px
-  &::-webkit-scrollbar-thumb
-    background fade-out(#fff, 50%)
-    border-radius 2px
-
-  yi-card
-    margin 1rem auto
 
 // responsive
 @import 'reports-mobile'
